feat(navigation): support stable decimal ordering for CMS items
- Change navigation order storage to decimal
- Heal duplicate static orders deterministically
- Preserve int.MaxValue as the unset-order sentinel
- Resolve generated unset orders near their siblings
- Use depth-based spacing for healed orders
- Add navigation ordering tests
- Update built-in navigation items with spaced orders

+++ src/core/ComponentModel/Navigation/NavigationItemAttribute.cs
@@ -43,7 +43,7 @@ public abstract class NavigationItemAttribute : Attribute
     /// <summary>
     /// Gets / sets the order (only) among its siblings.
     /// </summary>
-    public int Order { get; set; }
+    public decimal Order { get; set; }
 
     /// <summary>
     /// Url of this navigation item, should be null for menu
